rag-pdf-qa/requirements.txt
fastapi
uvicorn
requests
pdfplumber
transformers
torch
pinecone-client
